requirements.txt
tweepy>=3.5.0
hazm>=0.5.2
sqlalchemy>=1.1.9
pymysql>=0.7.11
python-dateutil
